requirements.txt
fastapi
uvicorn[standard]
asyncpg
pydantic-settings
pyjwt
httpx
pytest
redis